requirements.txt
django==2.0.2   
django-bootstrap4==0.0.6 
django-dirtyfields==1.3.1 
django-filter==1.1.0   
django-model-utils==3.1.1   
djangorestframework==3.7.7   
docopt==0.6.2 
gunicorn==19.7.1   
psycopg2-binary==2.7.4                                                     
pytz==2018.3                 
